Adding a byte count to Lang::String, for keeping arbitrary data.
The byte count does not replace the trailing null byte, but serves as a complement.

The byte count is with respect to the _data_ in the string, and does not count the trailing null byte.

A predicate for valid UTF-8 content was also added, although, at the moment, only strings with valid UTF-8 content are possible to construct.