Additional string construction syntax, for computer-generated strings and raw data.
Please refer to the documentation (to be part of the following commit) for a description of the syntax.

As an example, here's a string that ends with a newline:
  "{{Here comes the end of the line!}A0}
There are also named characters in escape mode, all named by lower case letters:
  "{{Here comes the end of the line!}n}

Note that this means that
  "{n}
is a fairly readable and compact way to enter the newline character.