New feature: Prelude files.
Before any other input is read, each directory in the need file search path is searched for a file named "Shapes-Prelude".  If present, this file should be an ordinary Shapes extension file, except that it does not have the ".shext" suffix, and cannot be loaded via the ##needs scanner directive.  The content of the Shapes-Prelude file should be a sequence of ##needs directives.

By using ##needs directives in the Shapes-Prelude files, is it ensured that the user can shadow the default prelude definitions.  That is, even if the Shapes-Prelude in a "shadowed directory" late in the need search path is read, the ##needs directives therein will refer to the files that the user has provided in a directory earlier on the need search path.

All Shapes-Prelude files are inserted in an environment under the core bindings and parent to the user's top level bindings.  Hence, the user can shadow prelude bindings on the top level.