Doc: Hyperlinks for environment variables.
At the moment, the env-var[@name] tag does no formatting, but only links to the correct spot in the man page.  This sould be sufficient, since any variable that is of interest should be mentioned in the man page; other environment variables should not appear anywhere in the documentation.  Hence, there is no need to keep formatting and linking apart.