Improved indentation, added todos, bugs and cleaned up.
The indentation algorithm now also handles function definitions (based
on a not so elegent looking-back regex-based method, in addition to the
normal parse-partial-sexp business.  Ideally, syntax-ppss should be
extended to handle all kinds of Shapes expressions, in which case the
code would be much clearer, since it is normally a matter of inserting
a basic indent as soon as you are inside an expression.

The clarity of the code has detoriated, I need to think more on how to structure
it. Also need to think about how to handle operators.