Added flymake-mode support.
Uses the same command as for compilation, and also the same regular expression
for parsing the compiler output.

It seems like flymake does not make use of column numbers, which would be a
shame.  I will investigate this further, and possibly implement it myself.

Resource consumption could be decreased if the Shapes compiler supported a mode
where it did analysis only, no code generation.